NN model can allocate tensors

+++ PCAP_Firmware/src/pcap04_defs.h
@@ -81,7 +81,7 @@ extern "C" {
  * and offset corrections for all sensors on a single PCAP04 chip.
  */
 typedef struct {
-    uint32_t raw[NUM_SENSORS_PER_CHIP];     ///< Raw 24-bit measurement values from ADC
+    float raw[NUM_SENSORS_PER_CHIP];        ///< Raw 24-bit measurement values from ADC
     float final_val[NUM_SENSORS_PER_CHIP];  ///< Calibrated/processed measurement values
     float offset[NUM_SENSORS_PER_CHIP];     ///< Offset correction values for calibration
 } pcap_data_t;
